addresses are 9 bits not 8

+++ tos/chips/fm25lb/Fm25lb.h
@@ -5,9 +5,7 @@
 #ifndef __FM25LB_H__
 #define __FM25LB_H__
 
-//#include "Storage.h"
-
-typedef uint8_t fm25lb_addr_t;
-typedef uint8_t fm25lb_len_t;
+typedef uint16_t fm25lb_addr_t;
+typedef uint16_t fm25lb_len_t;
 
 #endif
